refactor: break applyRules to separate file to ease testing
